feat(virtual-background): Resize and drag and drop on virtual background.
