+++ react/features/virtual-background/functions.js
@@ -140,3 +140,88 @@ export function timeout(milliseconds: number, promise: Promise<*>): Promise<Obje
         promise.then(resolve, reject);
     });
 }
+
+// Define several math function.
+
+/**
+ * Get client rect.
+ *
+ * @param {Object} rotatedBox - Object that contains dimensions positions and rotation of resizable element.
+ * @returns {void}
+ */
+export function getClientRect(rotatedBox: Object) {
+    const { x, y, width, height } = rotatedBox;
+    const rad = rotatedBox.rotation;
+
+    const p1 = getCorner(x, y, 0, 0, rad);
+    const p2 = getCorner(x, y, width, 0, rad);
+    const p3 = getCorner(x, y, width, height, rad);
+    const p4 = getCorner(x, y, 0, height, rad);
+
+    const minX = Math.min(p1.x, p2.x, p3.x, p4.x);
+    const minY = Math.min(p1.y, p2.y, p3.y, p4.y);
+    const maxX = Math.max(p1.x, p2.x, p3.x, p4.x);
+    const maxY = Math.max(p1.y, p2.y, p3.y, p4.y);
+
+    return {
+        x: minX,
+        y: minY,
+        width: maxX - minX,
+        height: maxY - minY
+    };
+}
+
+/* eslint-disable max-params */
+/**
+ * Get Corner.
+ *
+ * @param {number} pivotX - The min x position.
+ * @param {number} pivotY - The min y position.
+ * @param {number} diffX - The box width.
+ * @param {number} diffY - The box height.
+ * @param {number} angle - Rotation details.
+ * @returns {void}
+ */
+export function getCorner(pivotX: number, pivotY: number, diffX: number, diffY: number, angle: number) {
+    const distance = Math.sqrt((diffX * diffX) + (diffY * diffY));
+
+    // Find angle from pivot to corner.
+    // eslint-disable-next-line no-param-reassign
+    angle += Math.atan2(diffY, diffX);
+
+    // Get new x and y and round it off to integer.
+    const x = pivotX + (distance * Math.cos(angle));
+    const y = pivotY + (distance * Math.sin(angle));
+
+    return {
+        x,
+        y
+    };
+}
+
+/**
+ * Get total boxes on drag action.
+ *
+ * @param {Object} boxes - Object that contains dimensions and positions of draggable element.
+ * @returns {void}
+ */
+export function getTotalBox(boxes: Object) {
+    let minX = Infinity;
+    let minY = Infinity;
+    let maxX = -Infinity;
+    let maxY = -Infinity;
+
+    boxes.forEach(box => {
+        minX = Math.min(minX, box.x);
+        minY = Math.min(minY, box.y);
+        maxX = Math.max(maxX, box.x + box.width);
+        maxY = Math.max(maxY, box.y + box.height);
+    });
+
+    return {
+        x: minX,
+        y: minY,
+        width: maxX - minX,
+        height: maxY - minY
+    };
+}
